three lists dnd working
