+++ frontend/src/courses.js
@@ -1,0 +1,28 @@
+const courses = [
+  {
+  id: 'ECE20001',
+  content: 'Linear Circuit Analysis I'
+  },
+  {
+  id: 'ECE20002',
+  content: 'Linear Circuit Analysis II'
+  },
+  {
+  id: 'ECE27000',
+  content: 'Digital Systems Design'
+  },
+  {
+  id: 'ECE36200',
+  content: 'Microprocessor Systems and Interfacing'
+  },
+  {
+  id: 'ECE36900',
+  content: 'Discrete Mathematics'
+  },
+  {
+  id: 'ECE40400',
+  content: 'Introduction to Computer Security',
+  }
+];
+
+export default courses;
