--- frontend/src/multiple-list-dnd.component.js
@@ -1,191 +1,0 @@
-import React, { Component } from 'react';
-import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
-
-const getItems = (count, offset = 0) =>[
-  {
-  id: 'ECE20001',
-  content: 'Linear Circuit Analysis I'
-  },
-  {
-  id: 'ECE20002',
-  content: 'Linear Circuit Analysis II'
-  },
-  {
-  id: 'ECE27000',
-  content: 'Digital Systems Design'
-  },
-  {
-  id: 'ECE36200',
-  content: 'Microprocessor Systems and Interfacing'
-  },
-  {
-  id: 'ECE36900',
-  content: 'Discrete Mathematics'
-  },
-  {
-  id: 'ECE40400',
-  content: 'Introduction to Computer Security',
-  }
-];
-
-const reorder = (list, startIndex, endIndex) => {
-  const result = Array.from(list);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-
-  return result;
-};
-
-// Move item from one list to other
-const move = (source, destination, droppableSource, droppableDestination) => {
-  const sourceClone = Array.from(source);
-  const destClone = Array.from(destination);
-  const [removed] = sourceClone.splice(droppableSource.index, 1);
-
-  destClone.splice(droppableDestination.index, 0, removed);
-
-  const result = {};
-  result[droppableSource.droppableId] = sourceClone;
-  result[droppableDestination.droppableId] = destClone;
-
-  return result;
-};
-
-const grid = 10;
-
-const getItemStyle = (isDragging, draggableStyle) => ({
-  userSelect: 'none',
-  padding: grid * 2,
-  margin: `0 0 ${grid}px 0`,
-
-  background: isDragging ? 'lightgreen' : 'white',
-
-  ...draggableStyle
-});
-
-const getListStyle = isDraggingOver => ({
-  background: isDraggingOver ? 'lightblue' : 'lightgrey',
-  padding: grid,
-  width: 250
-});
-
-class MultipleDragList extends Component {
-  state = {
-    items: getItems(3),
-    selected: []
-  };
-
-  // Defining unique ID for multiple lists
-  id2List = {
-    droppable: 'items',
-    droppable2: 'selected'
-  };
-
-  getList = id => this.state[this.id2List[id]];
-
-  onDragEnd = result => {
-    const { source, destination } = result;
-
-    if (!destination) {
-      return;
-    }
-
-    // Sorting in same list
-    if (source.droppableId === destination.droppableId) {
-      const items = reorder(
-        this.getList(source.droppableId),
-        source.index,
-        destination.index
-      );
-
-      let state = { items };
-
-      if (source.droppableId === 'droppable2') {
-        state = { selected: items };
-      }
-
-      this.setState(state);
-    }
-    // Interlist movement
-    else {
-      const result = move(
-        this.getList(source.droppableId),
-        this.getList(destination.droppableId),
-        source,
-        destination
-      );
-
-      this.setState({
-        items: result.droppable,
-        selected: result.droppable2
-      });
-    }
-  };
-
-  render() {
-    return (
-      <div style={{ 'display': 'flex' }}>
-        <DragDropContext onDragEnd={this.onDragEnd}>
-          <Droppable droppableId="droppable">
-            {(provided, snapshot) => (
-              <div
-                ref={provided.innerRef}
-                style={getListStyle(snapshot.isDraggingOver)}>
-                {this.state.items.map((item, index) => (
-                  <Draggable
-                    key={item.id}
-                    draggableId={item.id}
-                    index={index}>
-                    {(provided, snapshot) => (
-                      <div
-                        ref={provided.innerRef}
-                        {...provided.draggableProps}
-                        {...provided.dragHandleProps}
-                        style={getItemStyle(
-                          snapshot.isDragging,
-                          provided.draggableProps.style
-                        )}>
-                        {item.content}
-                      </div>
-                    )}
-                  </Draggable>
-                ))}
-                {provided.placeholder}
-              </div>
-            )}
-          </Droppable>
-          <Droppable droppableId="droppable2">
-            {(provided, snapshot) => (
-              <div
-                ref={provided.innerRef}
-                style={getListStyle(snapshot.isDraggingOver)}>
-                {this.state.selected.map((item, index) => (
-                  <Draggable
-                    key={item.id}
-                    draggableId={item.id}
-                    index={index}>
-                    {(provided, snapshot) => (
-                      <div
-                        ref={provided.innerRef}
-                        {...provided.draggableProps}
-                        {...provided.dragHandleProps}
-                        style={getItemStyle(
-                          snapshot.isDragging,
-                          provided.draggableProps.style
-                        )}>
-                        {item.content}
-                      </div>
-                    )}
-                  </Draggable>
-                ))}
-                {provided.placeholder}
-              </div>
-            )}
-          </Droppable>
-        </DragDropContext>
-      </div>
-    );
-  }
-}
-
-export default MultipleDragList;
